+++ src/services/tasksService.ts
@@ -2,35 +2,34 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as path from 'path';
 
-import { Task, Gres, LogFile, TaskProperties } from '../models';
+import { BaseTask, loadObj, TaskArray, TaskObjTypes } from '../models';
 import { executeCmd } from '../utils/utils';
 
-async function getFinishTime(taskId: number[]): Promise<Map<number, string>> {
-    if (taskId.length === 0) { return new Map(); }
-    const jobid_len = 40;
 
-    const taskIdStr = taskId.map((v) => String(v));
-    const [out, err] = await executeCmd(`sacct -j ${taskId.join(',')} --format=JobID%${jobid_len},End%25`);
+async function getFinishTime(taskId: string[]): Promise<Map<string, string>> {
+    if (taskId.length === 0) { return new Map(); }
+
+    const [out, err] = await executeCmd(`sacct --array -n -X -P -j ${taskId.join(',')} --format=JobID,End`);
     if (err) {
         vscode.window.showErrorMessage(err);
         return new Map();
     }
     const lines = out.split('\n');
 
-    const finishTimes: Map<number, string> = new Map();
-    for (let j = 2; j < lines.length; j++) {
-        const line = lines[j];
-        const jobid = line.slice(0, jobid_len).trim();
-        if (taskIdStr.includes(jobid)) {
-            finishTimes.set(parseInt(jobid), line.slice(jobid_len + 1).trim());
+    const finishTimes: Map<string, string> = new Map();
+    lines.forEach(line => {
+        const [jobid, time] = line.split('|');
+        if (time !== 'Unknown') {
+            finishTimes.set(jobid, time);
         }
-    }
+    });
     return finishTimes;
 }
 
+
 export class TaskService {
     private storagePath;
-    private taskMap!: Map<number, Task>;
+    private taskMap!: Map<number, BaseTask>;
 
     private static _instance: TaskService | null = null;
     private constructor(context: vscode.ExtensionContext) {
@@ -56,72 +55,70 @@ export class TaskService {
         if (fs.existsSync(this.storagePath)) {
             const jsonData = fs.readFileSync(this.storagePath, 'utf8');
             const saveMap = JSON.parse(jsonData);
-            const taskMap = new Map(saveMap) as Map<number, TaskProperties>;
+            const taskMap = new Map(saveMap) as Map<number, TaskObjTypes>;
 
-            taskMap.forEach((v, k) => this.taskMap.set(k, Task.fromObject(v)));
+            taskMap.forEach((v, k) => this.taskMap.set(k, loadObj(v)));
         }
     }
 
     private async saveTask() {
-        const arrData = Array.from(this.taskMap.entries());
-        const jsonData = JSON.stringify(arrData, (k, v) => {
-            if (k === 'gres') {
-                if (v === null) { return 'N/A'; }
-                return Gres.prototype.toString.call(v);
-            }
-            if (k === 'out_path' || k === 'err_path' || k === 'command') {
-                return LogFile.prototype.toString.call(v);
-            }
-            return v;
-        });
+        const arrData = Array.from(this.taskMap.entries()).map(([key, task]) => ([
+            key,
+            task.toObj()
+        ]));
+        const jsonData = JSON.stringify(arrData);
         fs.writeFileSync(this.storagePath, jsonData, 'utf8');
     }
 
-    private async finishTask(taskId: number[]) {
+    private async finishTask(taskId: string[]) {
         if (taskId.length === 0) { return; }
         const endmap = await getFinishTime(taskId);
-        endmap.forEach((v, k) => this.taskMap.get(k)?.finish(v));
+        endmap.forEach((v, k) => {
+            const [jobid, arrid] = k.split('_');
+            this.taskMap.get(parseInt(jobid))?.finish(arrid !== undefined ? { [arrid]: v } : v);
+        });
     }
 
-    public addTask(...tasks: Task[]) {
+    public addTask(...tasks: BaseTask[]) {
         tasks.forEach(value => this.taskMap.set(value.jobid, value));
         this.saveTask();
     }
 
-    public async updateTask(...tasks: Task[]) {
+    public async updateTask(...tasks: BaseTask[]) {
         this.loadTask();
         const newId = tasks.map(value => value.jobid);
         const oldId = [...this.taskMap.values()].filter(v => !v.finished).map(v => v.jobid);
 
         const updateId = newId.filter(value => oldId.includes(value));
         const addId = newId.filter(value => !updateId.includes(value));
-        const dropId = oldId.filter(value => !updateId.includes(value));
-
 
         tasks.filter(value => addId.includes(value.jobid))
             .forEach(value => this.taskMap.set(value.jobid, value));
         tasks.filter(value => updateId.includes(value.jobid))
             .forEach(value => this.taskMap.get(value.jobid)?.update(value));
-        await this.finishTask(dropId);
+
+        const newArrId = tasks.map(value => value.jobArrayId).flat();
+        const oldArrId = [...this.taskMap.values()].filter(v => !v.finished).map(v => v.jobArrayId).flat();
+
+        const updateArrId = newArrId.filter(value => oldArrId.includes(value));
+        const dropArrId = oldArrId.filter(value => !updateArrId.includes(value));
+
+        await this.finishTask(dropArrId);
 
         this.saveTask();
     }
 
-    public deleteTask(...tasks: number[]): void;
-    public deleteTask(...tasks: Task[]): void;
-    public deleteTask(...tasks: Task[] | number[]): void {
-        if (tasks.length === 0) {
-            return;
-        }
-        if (typeof tasks[0] === "number") {
-            tasks.forEach(value => this.taskMap.delete(value as number));
+    public deleteTask(task: BaseTask | number, arrayId?: number | null): void {
+        const taskId = typeof task === "number" ? task : task.jobid;
+        if (arrayId === undefined || arrayId === null) {
+            this.taskMap.delete(taskId);
         } else {
-            tasks.forEach(value => this.taskMap.delete((value as Task).jobid));
+            delete (this.taskMap.get(taskId) as TaskArray).subTasks[arrayId];
         }
         this.saveTask();
     }
 
-    public getTask(): Task[] {
+    public getTask(): BaseTask[] {
         return [...this.taskMap.values()];
     }