+++ src/view/components/nodeItem.ts
@@ -4,7 +4,7 @@ import { Node, NodeState } from '../../models';
 function nodeDescription(node: Node): string {
     if (node.state === NodeState.ALLOC) { return vscode.l10n.t(`Node has been allocated`); }
     if (node.state === NodeState.IDLE || node.state === NodeState.MIXED) {
-        return `${node.gres ?? 'No GRES'}\t${node.allocMemory}GB/${node.memory}GB`;
+        return `${node.gres ?? 'No GRES'} \t${Math.round((node.memory - node.allocMemory) * 1000) / 1000}GB / ${node.memory}GB \t${node.idleCpu} / ${node.cpu} CPUs`;
     }
     return vscode.l10n.t(`Node not available due to state "{0}"`, node.state);
 }
@@ -24,7 +24,11 @@ export class NodeItem extends vscode.TreeItem {
     constructor(public readonly node: Node) {
         super(node.nodeid, vscode.TreeItemCollapsibleState.None);
         this.description = nodeDescription(node);
-        this.tooltip = vscode.l10n.t(`{0} state: {1}`, node.nodeid, node.state);
+        this.tooltip = new vscode.MarkdownString(`
+| nodeid | state | GRES (idle/total) | memory (i/t) | CPUs (i/t) |
+|:--:|:--:|:--:|:--:|:--:|
+| ${node.nodeid} | ${node.state} | ${node.gres ?? 'No GRES'} | ${Math.round((node.memory - node.allocMemory) * 1000) / 1000}GB / ${node.memory}GB | ${node.idleCpu} / ${node.cpu} |
+            `);
         this.iconPath = nodeIcon(node);
         this.contextValue = (node.state === NodeState.IDLE || node.state === NodeState.MIXED) ? 'nodeItem' : 'unavailNodeItem';
     }